+++ bin/node/src/commands/bundled.rs
@@ -1,6 +1,7 @@
 use std::collections::HashMap;
 use std::path::PathBuf;
 use std::sync::Arc;
+use std::time::Duration;
 
 use anyhow::Context;
 use miden_node_block_producer::BlockProducer;
@@ -16,9 +17,11 @@ use url::Url;
 use super::{ENV_DATA_DIRECTORY, ENV_RPC_URL};
 use crate::commands::{
     BlockProducerConfig,
+    DEFAULT_TIMEOUT,
     ENV_ENABLE_OTEL,
     ENV_GENESIS_CONFIG_FILE,
     NtxBuilderConfig,
+    duration_to_human_readable_string,
 };
 
 #[derive(clap::Subcommand)]
@@ -67,6 +70,17 @@ pub enum BundledCommand {
         /// OpenTelemetry documentation. See our operator manual for further details.
         #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
         enable_otel: bool,
+
+        /// Maximum duration a gRPC request is allocated before being dropped by the server.
+        ///
+        /// This may occur if the server is overloaded or due to an internal bug.
+        #[arg(
+            long = "grpc.timeout",
+            default_value = &duration_to_human_readable_string(DEFAULT_TIMEOUT),
+            value_parser = humantime::parse_duration,
+            value_name = "DURATION"
+        )]
+        grpc_timeout: Duration,
     },
 }
 
@@ -94,7 +108,11 @@ impl BundledCommand {
                 block_producer,
                 ntx_builder,
                 enable_otel: _,
-            } => Self::start(rpc_url, data_directory, ntx_builder, block_producer).await,
+                grpc_timeout,
+            } => {
+                Self::start(rpc_url, data_directory, ntx_builder, block_producer, grpc_timeout)
+                    .await
+            },
         }
     }
 
@@ -104,6 +122,7 @@ impl BundledCommand {
         data_directory: PathBuf,
         ntx_builder: NtxBuilderConfig,
         block_producer: BlockProducerConfig,
+        grpc_timeout: Duration,
     ) -> anyhow::Result<()> {
         let should_start_ntb = !ntx_builder.disabled;
         // Start listening on all gRPC urls so that inter-component connections can be created
@@ -152,6 +171,7 @@ impl BundledCommand {
                     block_producer_listener: store_block_producer_listener,
                     ntx_builder_listener: store_ntx_builder_listener,
                     data_directory: data_directory_clone,
+                    grpc_timeout,
                 }
                 .serve()
                 .await
@@ -182,6 +202,7 @@ impl BundledCommand {
                         max_batches_per_block: block_producer.max_batches_per_block,
                         max_txs_per_batch: block_producer.max_txs_per_batch,
                         production_checkpoint: checkpoint,
+                        grpc_timeout,
                     }
                     .serve()
                     .await
@@ -197,6 +218,7 @@ impl BundledCommand {
                     listener: grpc_rpc,
                     store: store_rpc_address,
                     block_producer: Some(block_producer_address),
+                    grpc_timeout,
                 }
                 .serve()
                 .await