+++ bin/node/src/commands/store.rs
@@ -1,4 +1,5 @@
 use std::path::{Path, PathBuf};
+use std::time::Duration;
 
 use anyhow::Context;
 use miden_node_store::Store;
@@ -12,7 +13,12 @@ use super::{
     ENV_STORE_NTX_BUILDER_URL,
     ENV_STORE_RPC_URL,
 };
-use crate::commands::{ENV_ENABLE_OTEL, ENV_GENESIS_CONFIG_FILE};
+use crate::commands::{
+    DEFAULT_TIMEOUT,
+    ENV_ENABLE_OTEL,
+    ENV_GENESIS_CONFIG_FILE,
+    duration_to_human_readable_string,
+};
 
 #[allow(clippy::large_enum_variant, reason = "single use enum")]
 #[derive(clap::Subcommand)]
@@ -62,6 +68,17 @@ pub enum StoreCommand {
         /// OpenTelemetry documentation. See our operator manual for further details.
         #[arg(long = "enable-otel", default_value_t = false, env = ENV_ENABLE_OTEL, value_name = "BOOL")]
         enable_otel: bool,
+
+        /// Maximum duration a gRPC request is allocated before being dropped by the server.
+        ///
+        /// This may occur if the server is overloaded or due to an internal bug.
+        #[arg(
+            long = "grpc.timeout",
+            default_value = &duration_to_human_readable_string(DEFAULT_TIMEOUT),
+            value_parser = humantime::parse_duration,
+            value_name = "DURATION"
+        )]
+        grpc_timeout: Duration,
     },
 }
 
@@ -82,7 +99,17 @@ impl StoreCommand {
                 block_producer_url,
                 data_directory,
                 enable_otel: _,
-            } => Self::start(rpc_url, ntx_builder_url, block_producer_url, data_directory).await,
+                grpc_timeout,
+            } => {
+                Self::start(
+                    rpc_url,
+                    ntx_builder_url,
+                    block_producer_url,
+                    data_directory,
+                    grpc_timeout,
+                )
+                .await
+            },
         }
     }
 
@@ -99,6 +126,7 @@ impl StoreCommand {
         ntx_builder_url: Url,
         block_producer_url: Url,
         data_directory: PathBuf,
+        grpc_timeout: Duration,
     ) -> anyhow::Result<()> {
         let rpc_listener = rpc_url
             .to_socket()
@@ -126,6 +154,7 @@ impl StoreCommand {
             ntx_builder_listener,
             block_producer_listener,
             data_directory,
+            grpc_timeout,
         }
         .serve()
         .await