+++ crates/rpc/src/tests.rs
@@ -136,6 +136,7 @@ async fn rpc_startup_is_robust_to_network_failures() {
             ntx_builder_listener,
             block_producer_listener,
             data_directory: data_directory.path().to_path_buf(),
+            grpc_timeout: Duration::from_secs(10),
         }
         .serve()
         .await
@@ -287,6 +288,7 @@ async fn start_rpc() -> (ApiClient, std::net::SocketAddr, std::net::SocketAddr) 
             listener: rpc_listener,
             store: store_addr,
             block_producer: Some(block_producer_addr),
+            grpc_timeout: Duration::from_secs(30),
         }
         .serve()
         .await
@@ -323,6 +325,7 @@ async fn start_store(store_addr: SocketAddr) -> (Runtime, TempDir) {
             ntx_builder_listener,
             block_producer_listener,
             data_directory: dir,
+            grpc_timeout: Duration::from_secs(30),
         }
         .serve()
         .await